millisecond millisecond millisecond millisecond

+++ include/knot/engine.h
@@ -32,12 +32,12 @@ struct TimerTask {
     /**
      * @brief Time remaining until the next event trigger, in milliseconds.
      */
-    double time;
+    int time;
 
     /**
      * @brief The interval between repeated triggers, in milliseconds.
      */
-    double interval;
+    int interval;
 
     /**
      * @brief The unique identifier of the timer task.
@@ -63,7 +63,7 @@ public:
      *
      * @param deltaTime The elapsed time since the previous frame, in seconds.
      */
-    using RenderLoopCallback = std::function<void(double)>;
+    using RenderLoopCallback = std::function<void(float)>;
 
     Engine() = default;
 
@@ -226,7 +226,7 @@ public:
      *
      * @return The unique ID assigned to the timer task.
      */
-    int after(double interval, uint32_t eventCode, std::any eventData = {});
+    int after(int interval, uint32_t eventCode, std::any eventData = {});
 
     /**
      * @brief Schedules a repeating user event.
@@ -240,7 +240,7 @@ public:
      *
      * @return The unique ID assigned to the timer task.
      */
-    int repeat(double interval, uint32_t eventCode, std::any eventData = {});
+    int repeat(int interval, uint32_t eventCode, std::any eventData = {});
 
     /**
      * @brief Requests the engine to stop running.
@@ -275,8 +275,8 @@ private:
     bool shouldQuit = false;
 
     bool initialized = false;
-    double deltaTime = 0.0f;
-    double lastFrame = 0.0f;
+    float deltaTime = 0.0f;
+    double lastFrame = 0.0;
 
     Scene* scene = nullptr;
